requirements.txt
Django==3.2.12
django-environ==0.8.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
django-rest-swagger==2.2.0
aiogram==2.20
